Require custom expiration time when generating JWT tokens

+++ src/main/java/pragmasoft/k1teauth/oauth/dto/TokenResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record TokenResponse(
         @JsonProperty("access_token") String token,
-        @JsonProperty("expires_in") int expiresInSecs,
+        @JsonProperty("expires_in") long expiresInSecs,
         @JsonProperty("token_type") String tokenType
 ) {
 }
